Improve conflict set calculation for sums.
Previous method just checked against the last cell, whereas now we
greedily check all cells to see if we can expand the current set.

Impact on killer sets (number of guesses):

 - `TAREK_ALL`: 44063 -> 44473
 - `EXTREME_KILLERS`: 61853 -> 68381
 - `MATHEMAGIC_KILLERS`: 9362564 -> 8245614

The regression on the first two are minor, while delivering a huge win
for the last.

Identified by `Vorados` on the CTC discord.