Move shortfall calculation into its own seperate backward pass.
The shortfall calculation now takes advantage of knowing exactly which
values are available for increasing visibility, as suggested by an
implementation by `Vorados` on the CTC discord.

The shortfall calculation has been moved into a backward pass
immediately after the forward passes:
 - There is little impact on the possible deductions by deferring this.
 - Allows us to keep the code clean by seperating out the distinct
   calculations.

It is still a seperate backward pass to the next one used to restrict
minimum values.