Fix bug where `lastContradictionCell` could be used before it is initialized.
`lastContradictionCell` was often being read from the wrong `recFrame`,
which was often not yet initialized for the frame. This would have
resulted in an unexpected cell being used. Most of the time this would
probably still have been a useful cell to enforce so didn't hurt
performance too much.

It did break the consistent of the back button in the UI, because that
restarts the solver using the same solver object and hence
`lastContradictionCell` would have still have its values from the last
run.

This never impacts the correctness of the generated solutions as
`lastContradictionCell` is suggesting extra constraints to be enforced.

Fixed by rearranging the `run` method so that
`recFrame.lastContradictionCell` is used before a new `recFrame` is
created so we are always reading from an initialized `recFrame`.

Impact on #guesses:
                        Before  After
    TAREK_ALL           33604   33145
    EXTREME_KILLERS     59876   55464
    HARD_THERMOS        92764   112255
    HARD_RENBAN         327574  334185
    MATHEMAGIC_KILLERS  2576535 2530035

Minor (often positive) impact, except noticable regression on `HARD_THERMOS`.